Allow the configuration of the Odoo main directory
For Odoo <= 9.0, the main directory was "openerp". On later
versions, the directory is called "odoo". The runbot module tries
to auto-detect it, but it fails if the Odoo repository is pulled
as a dependency rather than as the main repository.

This commit adds a field for manual configuration of the directory.